refactor: move cel package up
