+++ pkg/cel/libs/jwt/lib.go
@@ -6,7 +6,7 @@ import (
 	"github.com/google/cel-go/cel"
 	"github.com/google/cel-go/common/types"
 	"github.com/google/cel-go/ext"
-	"github.com/kyverno/kyverno-envoy-plugin/pkg/authz/cel/libs/jwk"
+	"github.com/kyverno/kyverno-envoy-plugin/pkg/cel/libs/jwk"
 )
 
 type lib struct{}
