+++ pkg/cel/variables.go
