+++ pkg/engine/apol/compiler/policy.go
@@ -8,8 +8,8 @@ import (
 	"github.com/google/cel-go/cel"
 	"github.com/google/cel-go/common/types"
 	"github.com/google/cel-go/common/types/ref"
-	authzcel "github.com/kyverno/kyverno-envoy-plugin/pkg/authz/cel"
-	"github.com/kyverno/kyverno-envoy-plugin/pkg/authz/cel/utils"
+	authzcel "github.com/kyverno/kyverno-envoy-plugin/pkg/cel"
+	"github.com/kyverno/kyverno-envoy-plugin/pkg/cel/utils"
 	"github.com/kyverno/kyverno-envoy-plugin/pkg/engine"
 	"go.uber.org/multierr"
 	admissionregistrationv1 "k8s.io/api/admissionregistration/v1"
